GH-5286 Generalize concepts for evaluation steps and condition
